move Calorie, Name & Serving classes into attribute package
