+++ src/main/java/seedu/address/model/recipe/attribute/Serving.java
@@ -1,4 +1,4 @@
-package seedu.address.model.recipe;
+package seedu.address.model.recipe.attribute;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
